remove tv-list-event to separate file
