+++ lib/domain/blocs/tv/tv_on_the_air_bloc.dart
@@ -5,6 +5,7 @@ import 'package:bloc_concurrency/bloc_concurrency.dart';
 // Project imports:
 import 'package:comics_db_app/configuration/configuration.dart';
 import 'package:comics_db_app/domain/api_client/movie_and_tv_api_client.dart';
+import 'package:comics_db_app/domain/blocs/tv/tv_list_event.dart';
 import 'package:comics_db_app/domain/blocs/tv/tv_list_state.dart';
 import 'package:comics_db_app/domain/blocs/tv/tv_popular_list_bloc.dart';
 import 'package:comics_db_app/domain/entity/popular_tv_response.dart';
