+++ lib/ui/widgets/tv_airing_today/tv_airing_today_cubit.dart
@@ -9,8 +9,8 @@ import 'package:bloc/bloc.dart';
 
 // Project imports:
 import 'package:comics_db_app/domain/blocs/tv/tv_airing_today_list_bloc.dart';
+import 'package:comics_db_app/domain/blocs/tv/tv_list_event.dart';
 import 'package:comics_db_app/domain/blocs/tv/tv_list_state.dart';
-import 'package:comics_db_app/domain/blocs/tv/tv_popular_list_bloc.dart';
 import 'package:comics_db_app/domain/entity/tv.dart';
 import 'package:comics_db_app/ui/navigation/main_navigation.dart';
 import 'package:comics_db_app/ui/widgets/tv_list/components/tv_list_data.dart';
