+++ lib/ui/widgets/tv_on_the_air/tv_on_the_air_cubit.dart
@@ -9,9 +9,9 @@ import 'package:bloc/bloc.dart';
 import 'package:intl/intl.dart';
 
 // Project imports:
+import 'package:comics_db_app/domain/blocs/tv/tv_list_event.dart';
 import 'package:comics_db_app/domain/blocs/tv/tv_list_state.dart';
 import 'package:comics_db_app/domain/blocs/tv/tv_on_the_air_bloc.dart';
-import 'package:comics_db_app/domain/blocs/tv/tv_popular_list_bloc.dart';
 import 'package:comics_db_app/domain/entity/tv.dart';
 import 'package:comics_db_app/ui/navigation/main_navigation.dart';
 import 'package:comics_db_app/ui/widgets/tv_list/components/tv_list_data.dart';
